Corrections alpha

+++ trunk/genealogie-parent/genealogie-model/src/main/java/org/maugtaurus/projects/genealogie/persistance/dao/person/PersonDao.java
@@ -5,5 +5,5 @@ import org.maugtaurus.projects.genealogie.persistance.model.person.Person;
 public interface PersonDao {
 	void save(Person person) throws Exception;
 	void delete(Person person) throws Exception;
-	Person getPersonById(int id);
+	Person getPersonById(long id);
 }
